update bootstrap to pull from upstream deps

+++ bootstrap.sh
@@ -1,44 +1,12 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
-# Function to convert dependencies to a valid environment variables
-sanitize_var_name() {
-    echo "$1" | tr '-' '_' | tr '[:lower:]' '[:upper:]'
-}
-
-# Read in dependencies.json file
-set_env_vars() {
-    local json_file="$1"
-
-    # Check if jq is installed
-    if ! command -v jq &> /dev/null; then
-        echo "Error: jq is not installed. Please install jq to parse JSON."
-        exit 1
-    fi
-
-    # Read each key-value pair from the JSON file
-    while IFS='=' read -r key value; do
-        # Sanitize the key to create a valid environment variable name
-        env_var=$(sanitize_var_name "$key")
-
-        # Set the environment variable
-        export "$env_var"="$value"
-        echo "Set $env_var=$value"
-    done < <(jq -r 'to_entries[] | .key + "=" + .value' "$json_file")
-}
-
-# Set environment variables from dependencies.json
-set_env_vars "dependencies.json"
-
-# Setting CERT_MANAGER causes issues when clusterctl init probes for it,
-# as unlike azimuth, we don't install it outselves
-unset CERT_MANAGER
-
-# Check a clouds.yaml file exists in the same directory as the script
-if [ ! -f clouds.yaml ]; then
-    echo "A clouds.yaml file is required in the same directory as this script"
-    exit 1
-fi
+echo "Installing required tools..."
+sudo apt-get install -y snapd python3-openstackclient
+export PATH=$PATH:/snap/bin
+sudo snap install kubectl --classic
+sudo snap install helm --classic
+sudo snap install yq
 
 echo "Updating system to apply latest security patches..."
 export DEBIAN_FRONTEND=noninteractive
@@ -49,12 +17,16 @@ sudo apt-get -o Dpkg::Options::="--force-confold" \
              -o Dpkg::Options::="--force-confdef" \
              -y -qq upgrade > /dev/null
 
-echo "Installing required tools..."
-sudo apt-get install -y snapd python3-openstackclient
-export PATH=$PATH:/snap/bin
-sudo snap install kubectl --classic
-sudo snap install helm --classic
-sudo snap install yq
+
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+# loading environments
+source "$SCRIPT_DIR"/set-env.sh
+
+# Check a clouds.yaml file exists in the same directory as the script
+if [ ! -f clouds.yaml ]; then
+    echo "A clouds.yaml file is required in the same directory as this script"
+    exit 1
+fi
 
 echo "Installing clusterctl..."
 curl --progress-bar -L "https://github.com/kubernetes-sigs/cluster-api/releases/download/${CLUSTER_API}/clusterctl-linux-amd64" -o clusterctl
@@ -87,6 +59,7 @@ echo "Backing up existing kubeconfig if it exists..."
 if [ -f "$HOME/.kube/config" ]; then 
     mv -v "$HOME/.kube/config" "$HOME/.kube/config.bak"
 fi
+
 sudo microk8s.config | sudo tee ~/.kube/config
 sudo chown "$USER" ~/.kube/config
 sudo chmod 600 ~/.kube/config
@@ -101,8 +74,10 @@ helm repo add capi https://azimuth-cloud.github.io/capi-helm-charts
 helm repo add capi-addons https://azimuth-cloud.github.io/cluster-api-addon-provider
 helm repo update
 helm upgrade cluster-api-addon-provider capi-addons/cluster-api-addon-provider --create-namespace --install --wait -n clusters --version "${ADDON_PROVIDER}"
-kubectl apply -f "https://github.com/k-orc/openstack-resource-controller/releases/download/${KORC}/install.yaml"
+kubectl apply -f "https://github.com/k-orc/openstack-resource-controller/releases/download/v${KORC}/install.yaml"
 
+echo ""
+echo "=============================================================================="
 echo "You are now ready to create a cluster following the remaining instructions..."
-
 echo "https://stfc.atlassian.net/wiki/spaces/CLOUDKB/pages/211878034/Cluster+API+Setup"
+echo "=============================================================================="
